requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
greenlet==1.1.2
idna==3.3
lxml==4.7.1
mysql-connector-python==8.0.27
protobuf==3.19.1
PyMySQL==1.0.2
requests==2.26.0
SQLAlchemy==1.4.29
urllib3==1.26.7
uWSGI==2.0.20
websockets==10.1
